Flatten test directory and unify one-file-per-module convention
Closes #381.

- Move test/syntax/*.py and test/semantics/*.py up to test/, so each
  discopy/<module>.py again has a matching test/<module>.py.
- Move test/utils/utils.py to test/utils.py and relocate the shared test
  data (pickles, zx-graph.json) to a new test/fixtures/ folder.
- Merge test/quantum/eval.py into test/quantum/circuit.py so the quantum
  submodule keeps one test file per module (Circuit.eval lives in circuit).
- Update pyproject.toml testpaths/ignore and the paths referenced from
  test/utils.py and test/quantum/zx.py, plus the AGENTS.md description.
