+++ test/balanced.py
