+++ test/fixtures/pickles/main/quantum.Qudit.pickle
