+++ test/fixtures/pickles/main/quantum.zx.Diagram.pickle
