+++ test/fixtures/pickles/main/rigid.Box.pickle
