+++ test/fixtures/pickles/rigid.Box.legacy.pickle
