+++ test/fixtures/pickles/src/rigid_Box.py
