+++ test/fixtures/pickles/src/tensor_Box.py
