+++ test/rigid.py
