+++ test/symmetric.py
